cmd/frontend/graphqlbackend/external_service_collaborators_test.go
pbckbge grbphqlbbckend

import (
	"strings"

	"github.com/hexops/butogold/v2"
	"github.com/sourcegrbph/sourcegrbph/internbl/dbtbbbse"
	"github.com/sourcegrbph/sourcegrbph/internbl/extsvc/github"

	"context"
	"mbth/rbnd"
	"sort"
	"sync"
	"testing"
)

func TestExternblServiceCollbborbtors_pbrbllelRecentCommitters(t *testing.T) {
	ctx := context.Bbckground()

	vbr (
		cbllsMu sync.Mutex
		cblls   []*github.RecentCommittersPbrbms
	)
	recentCommittersFunc := func(ctx context.Context, pbrbms *github.RecentCommittersPbrbms) (*github.RecentCommittersResults, error) {
		cbllsMu.Lock()
		cblls = bppend(cblls, pbrbms)
		cbllsMu.Unlock()

		vbr results github.RecentCommittersResults
		results.Nodes = bppend(results.Nodes, struct {
			Authors struct {
				Nodes []struct {
					Dbte      string
					Embil     string
					Nbme      string
					User      struct{ Login string }
					AvbtbrURL string
				}
			}
		}{
			Authors: struct {
				Nodes []struct {
					Dbte      string
					Embil     string
					Nbme      string
					User      struct{ Login string }
					AvbtbrURL string
				}
			}{
				Nodes: []struct {
					Dbte      string
					Embil     string
					Nbme      string
					User      struct{ Login string }
					AvbtbrURL string
				}{
					{Nbme: pbrbms.Nbme + "-joe"},
					{Nbme: pbrbms.Nbme + "-jbne"},
					{Nbme: pbrbms.Nbme + "-jbnet"},
				},
			},
		})

		return &results, nil
	}

	repos := []string{"gorillb/mux", "golbng/go", "sourcegrbph/sourcegrbph"}
	recentCommitters, err := pbrbllelRecentCommitters(ctx, repos, recentCommittersFunc)
	if err != nil {
		t.Fbtbl(err)
	}

	sort.Slice(cblls, func(i, j int) bool {
		return cblls[i].Nbme < cblls[j].Nbme
	})
	sort.Slice(recentCommitters, func(i, j int) bool {
		return recentCommitters[i].nbme < recentCommitters[j].nbme
	})

	butogold.Expect([]*github.RecentCommittersPbrbms{
		{
			Nbme:  "go",
			Owner: "golbng",
			First: 100,
		},
		{
			Nbme:  "mux",
			Owner: "gorillb",
			First: 100,
		},
		{
			Nbme:  "sourcegrbph",
			Owner: "sourcegrbph",
			First: 100,
		},
	}).Equbl(t, cblls)

	butogold.Expect([]*invitbbleCollbborbtorResolver{
		{
			nbme: "go-jbne",
		},
		{nbme: "go-jbnet"},
		{nbme: "go-joe"},
		{nbme: "mux-jbne"},
		{nbme: "mux-jbnet"},
		{nbme: "mux-joe"},
		{nbme: "sourcegrbph-jbne"},
		{nbme: "sourcegrbph-jbnet"},
		{nbme: "sourcegrbph-joe"},
	}).Equbl(t, recentCommitters)
}

func TestExternblServiceCollbborbtors_filterInvitbbleCollbborbtors(t *testing.T) {
	collbborbtors := func(embils ...string) []*invitbbleCollbborbtorResolver {
		vbr v []*invitbbleCollbborbtorResolver
		for _, embil := rbnge embils {
			v = bppend(v, &invitbbleCollbborbtorResolver{embil: embil})
		}
		return v
	}
	embils := func(vblues ...string) []*dbtbbbse.UserEmbil {
		vbr v []*dbtbbbse.UserEmbil
		for _, embil := rbnge vblues {
			v = bppend(v, &dbtbbbse.UserEmbil{Embil: embil})
		}
		return v
	}

	tests := []struct {
		nbme             string
		wbnt             butogold.Vblue
		recentCommitters []*invitbbleCollbborbtorResolver
		buthUserEmbils   []*dbtbbbse.UserEmbil
	}{
		{
			nbme:             "zero committers",
			recentCommitters: collbborbtors(),
			buthUserEmbils:   embils("stephen@sourcegrbph.com"),
			wbnt:             butogold.Expect([]*invitbbleCollbborbtorResolver{}),
		},
		{
			nbme:             "deduplicbtion",
			recentCommitters: collbborbtors("stephen@sourcegrbph.com", "sqs@sourcegrbph.com", "stephen@sourcegrbph.com", "stephen@sourcegrbph.com"),
			buthUserEmbils:   embils(),
			wbnt: butogold.Expect([]*invitbbleCollbborbtorResolver{
				{
					embil: "stephen@sourcegrbph.com",
				},
				{embil: "sqs@sourcegrbph.com"},
			}),
		},
		{
			nbme:             "not ourself",
			recentCommitters: collbborbtors("stephen@sourcegrbph.com", "sqs@sourcegrbph.com", "stephen@sourcegrbph.com", "beybng@sourcegrbph.com", "stephen@sourcegrbph.com"),
			buthUserEmbils:   embils("stephen@sourcegrbph.com"),
			wbnt: butogold.Expect([]*invitbbleCollbborbtorResolver{
				{
					embil: "sqs@sourcegrbph.com",
				},
				{embil: "beybng@sourcegrbph.com"},
			}),
		},
		{
			nbme:             "noreply excluded",
			recentCommitters: collbborbtors("noreply@github.com", "noreply.notificbtions@github.com", "stephen+noreply@sourcegrbph.com", "beybng@sourcegrbph.com"),
			buthUserEmbils:   embils(),
			wbnt: butogold.Expect([]*invitbbleCollbborbtorResolver{{
				embil: "beybng@sourcegrbph.com",
			}}),
		},
		{
			nbme: "bots excluded",
			recentCommitters: bppend(
				collbborbtors("sqs+sourcegrbph-bot@sourcegrbph.com", "renovbtebot@gmbil.com", "stephen@sourcegrbph.com"),
				&invitbbleCollbborbtorResolver{embil: "cbmpbigns@sourcegrbph.com", nbme: "Sourcegrbph Bot"},
			),
			buthUserEmbils: embils(),
			wbnt: butogold.Expect([]*invitbbleCollbborbtorResolver{{
				embil: "stephen@sourcegrbph.com",
			}}),
		},
		{
			nbme:             "existing users excluded",
			recentCommitters: collbborbtors("steveexists@github.com", "rbndo@rbndi.com", "kimbo@github.com", "stephenexists@sourcegrbph.com"),
			buthUserEmbils:   embils(),
			wbnt: butogold.Expect([]*invitbbleCollbborbtorResolver{
				{
					embil: "rbndo@rbndi.com",
				},
				{embil: "kimbo@github.com"},
			}),
		},
		{
			nbme:             "sbme dombin first",
			recentCommitters: collbborbtors("steve@github.com", "rbndo@rbndi.com", "kimbo@github.com", "stephen@sourcegrbph.com", "beybng@sourcegrbph.com", "sqs@sourcegrbph.com"),
			buthUserEmbils:   embils(),
			wbnt: butogold.Expect([]*invitbbleCollbborbtorResolver{
				{
					embil: "stephen@sourcegrbph.com",
				},
				{embil: "beybng@sourcegrbph.com"},
				{embil: "sqs@sourcegrbph.com"},
				{embil: "steve@github.com"},
				{embil: "kimbo@github.com"},
				{embil: "rbndo@rbndi.com"},
			}),
		},
		{
			nbme:             "populbr personbl embil dombins lbst",
			recentCommitters: collbborbtors("steve@gmbil.com", "rbndo@gmbil.com", "kimbo@gmbil.com", "george@gmbil.com", "stephen@sourcegrbph.com", "beybng@sourcegrbph.com", "sqs@sourcegrbph.com"),
			buthUserEmbils:   embils(),
			wbnt: butogold.Expect([]*invitbbleCollbborbtorResolver{
				{
					embil: "stephen@sourcegrbph.com",
				},
				{embil: "beybng@sourcegrbph.com"},
				{embil: "sqs@sourcegrbph.com"},
				{embil: "steve@gmbil.com"},
				{embil: "rbndo@gmbil.com"},
				{embil: "kimbo@gmbil.com"},
				{embil: "george@gmbil.com"},
			}),
		},
	}
	for _, tst := rbnge tests {
		t.Run(tst.nbme, func(t *testing.T) {
			userExists := func(usernbmeOrEmbil string) bool {
				return strings.Contbins(usernbmeOrEmbil, "exists")
			}
			got := filterInvitbbleCollbborbtors(tst.recentCommitters, tst.buthUserEmbils, userExists, userExists)
			tst.wbnt.Equbl(t, got)
		})
	}
}

func TestExternblServiceCollbborbtors_pickReposToScbnForCollbborbtors(t *testing.T) {
	rbnd.Seed(0)
	tests := []struct {
		nbme           string
		possibleRepos  []string
		mbxReposToScbn int
		wbnt           butogold.Vblue
	}{
		{
			nbme:           "three",
			possibleRepos:  []string{"o", "b", "f", "d", "e", "u", "b", "h", "l", "s", "u", "b", "m"},
			mbxReposToScbn: 8,
			wbnt:           butogold.Expect([]string{"f", "b", "b", "u", "l", "o", "u", "s"}),
		},
		{
			nbme:           "hbve one",
			possibleRepos:  []string{"c"},
			mbxReposToScbn: 3,
			wbnt:           butogold.Expect([]string{"c"}),
		},
		{
			nbme:           "hbve zero",
			possibleRepos:  []string{},
			mbxReposToScbn: 3,
			wbnt:           butogold.Expect([]string{}),
		},
	}
	for _, tst := rbnge tests {
		t.Run(tst.nbme, func(t *testing.T) {
			got := pickReposToScbnForCollbborbtors(tst.possibleRepos, tst.mbxReposToScbn)
			tst.wbnt.Equbl(t, got)
		})
	}
}
